+++ lib/Model/LinkFacet.php
@@ -8,10 +8,20 @@
 
 namespace SebLucas\Cops\Model;
 
-use SebLucas\Cops\Output\Format;
-
-class LinkFacet extends Link
+/**
+ * From https://specs.opds.io/opds-1.2#4-facets
+ * An Acquisition Feed MAY offer multiple links to reorder the Publications listed
+ * in the feed or limit them to a subset. This specification defines one new relation
+ * to identify such links as Facets:
+ *   http://opds-spec.org/facet: An Acquisition Feed with a subset or an alternate order
+ *   of the Publications listed.
+ *
+ * Links using this relation MUST only appear in Acquisition Feeds.
+ */
+class LinkFacet extends LinkAcquisition
 {
+    public const LINK_RELATION = "http://opds-spec.org/facet";
+
     /** @var string|null */
     public $facetGroup;
     public bool $activeFacet;
@@ -29,21 +39,10 @@ class LinkFacet extends Link
      */
     public function __construct($phref, $ptitle = null, $pfacetGroup = null, $pactiveFacet = false, $pthreadCount = null, $database = null)
     {
-        parent::__construct($phref, Link::OPDS_PAGING_TYPE, "http://opds-spec.org/facet", $ptitle);
-        $this->href = Format::addDatabaseParam($this->href, $database);
+        parent::__construct($phref, self::LINK_RELATION, $ptitle, $database);
+        //$this->href = Format::addDatabaseParam($this->href, $database);
         $this->facetGroup = $pfacetGroup;
         $this->activeFacet = $pactiveFacet;
         $this->threadCount = $pthreadCount;
-    }
-
-    /**
-     * Summary of hrefXhtml
-     * @param string $endpoint
-     * @return string
-     */
-    public function hrefXhtml($endpoint = '')
-    {
-        // LinkFacet()->href is relative to endpoint
-        return $endpoint . $this->href;
     }
 }